Add play choice buttons to UI
Added rock, paper, and scissors button to the UI. When a button is
clicked, the event is captured, the computer make it's move, and the
winner of the round is determined. The game still keeps score like
normal and a tied round results in replaying that round.

This change was made so that users do not have to type in their move
choice and instead, can click a button.